backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
psycopg2-binary>=2.9.9
httpx==0.26.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
